fix(gitlab): update ID types from int to int64 for API compatibility
- Change Group.ID and Project.ID from int to int64
- Update all function signatures accepting group/project IDs
- Update Config.GitlabGroupID and GitlabProjectID types
- Update all test files to use int64 for ID assertions
- Replace interface{} with any per linter recommendations

Adapts to gitlab.com/gitlab-org/api/client-go v1.3.0 which changed ID field types from int to int64.

+++ pkg/app/app.go
@@ -129,7 +129,7 @@ func (a *App) ExportGroup(ctx context.Context) error {
 }
 
 // ExportProject exports the project of the given ID.
-func (a *App) ExportProject(ctx context.Context, projectID int) error {
+func (a *App) ExportProject(ctx context.Context, projectID int64) error {
 	project, err := a.gitlabService.GetProject(ctx, projectID)
 	if err != nil {
 		return fmt.Errorf("failed to get project %d: %w", projectID, err)
